Generalized loading of score components from JSON
