+++ src/main/java/com/github/harverst/roper/model/common/BasicScoreComponent.java
@@ -3,6 +3,12 @@ package com.github.harverst.roper.model.common;
 import com.github.harverst.roper.model.ScoreComponent;
 import java.lang.Comparable;
 
+import org.json.JSONObject;
+import org.json.JSONException;
+
+import java.lang.Enum;
+import static java.lang.Enum.valueOf;
+
 /**
  * A ScoreComponent uses a component function to change the value of a score
  *
@@ -33,5 +39,39 @@ public abstract class BasicScoreComponent<P> implements ScoreComponent<P>
   {
     return phase;
   }
+  
+  // 
+  public static <P extends Enum<P> > ScoreComponent<P>
+    fromJson(Class<P> phaseType, JSONObject obj)
+  {
+    try
+    {
+      // Needs some sort of phaseType
+      int value = obj.getInt("value");
+      P phase = valueOf(phaseType, obj.getString("phase"));
+      switch(valueOf(BasicComponentType.class, obj.getString("type")))
+      {
+      case ADDITIVE:
+        return new AdditiveScoreComponent(phase, value);
+      case MULTIPLICATIVE:
+        return new MultiplicativeScoreComponent(phase, value);
+      case DIVISIVE:
+        return new DivisiveScoreComponent(phase, value);
+      default:
+        return null;
+      }
+    }
+    catch(JSONException e)
+    {
+      return null;
+    }
+  }
+}
+
+enum BasicComponentType
+{
+  ADDITIVE,
+  MULTIPLICATIVE,
+  DIVISIVE
 }
