+++ src/main/java/com/github/harverst/roper/model/fifthedition/FifthEditionCharacterComponent.java
@@ -3,6 +3,7 @@ package com.github.harverst.roper.model.fifthedition;
 import com.github.harverst.roper.model.Character;
 import com.github.harverst.roper.model.ScoreComponent;
 import com.github.harverst.roper.model.CharacterComponent;
+import com.github.harverst.roper.model.common.BasicScoreComponent;
 import com.github.harverst.roper.model.fifthedition.FifthEditionScoreType;
 import com.github.harverst.roper.model.fifthedition.FifthEditionPhase;
 import org.json.JSONObject;
@@ -20,12 +21,19 @@ import java.io.FileReader;
 public abstract class FifthEditionCharacterComponent
   implements CharacterComponent<FifthEditionScoreType, FifthEditionPhase>
 {
-  public static FifthEditionCharacterComponent loadFromFile(String filename)
+  public static CharacterComponent<FifthEditionScoreType, FifthEditionPhase>
+    fromJson(String source)
+    throws JSONException
   {
-    // 
-    JSONObject loaded = null;
-    // grab a new filename
-    return new FifthEditionCharacterJsonComponent(loaded);
+    JSONObject obj = new JSONObject(source);
+    return new FifthEditionCharacterJsonComponent(obj);
+  }
+  public static CharacterComponent<FifthEditionScoreType, FifthEditionPhase> 
+    fromJsonFile(String filename)
+    throws JSONException
+  {
+    // Open the json file and read all the contents
+    return fromJson(filename);
   }
 }
 
@@ -48,7 +56,8 @@ class FifthEditionCharacterJsonComponent extends FifthEditionCharacterComponent
         {
           try
           {
-            subList.add(FifthEditionScoreComponent.fromJson(
+            subList.add(BasicScoreComponent.fromJson(
+              FifthEditionPhase.class,
               subjson.getJSONObject(name)));
           }
           catch(JSONException e)
@@ -67,6 +76,7 @@ class FifthEditionCharacterJsonComponent extends FifthEditionCharacterComponent
   }
   public void composite(Character<FifthEditionScoreType, FifthEditionPhase> c)
   {
+    
   }
   public void seperate(Character<FifthEditionScoreType, FifthEditionPhase> c)
   {
