--- src/test/java/com/github/harverst/roper/model/fifthedition/FithEditionHitDice.java
@@ -1,7 +1,0 @@
-package com.github.harverst.roper.model.fifthedition;
-
-/**
- * Created by eli on 7/5/15.
- */
-public class FithEditionHitDice {
-}
